# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WhuzPay Payment Gateway Aggregator API
  description: |
    REST API for the WhuzPay QRIS payment gateway aggregator. Orchestrates
    payment providers (Cashi in production, an in-process mock for
    merchant sandbox) behind one consistent API and dashboard.

    Generated from the actual route table in `cmd/api/main.go` and the
    handler/service/domain source — not hand-guessed. See
    [docs/payment-state-machine.md](payment-state-machine.md) for the
    payment status lifecycle these endpoints read and write.
  version: "1.0"
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
security: []
paths:
...
  /admin/merchants:
    get:
      tags: [Admin - Merchants]
      summary: List merchants
      security: [{ adminBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive] } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedMerchants' } } } }
    post:
      tags: [Admin - Merchants]
      summary: Create a merchant
      security: [{ adminBearer: [] }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateMerchantRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/MerchantResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: Merchant with this email already exists }
...
  /admin/merchants/{id}:
    get:
      tags: [Admin - Merchants]
      summary: Get merchant
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/PathID' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MerchantResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Admin - Merchants]
      summary: Update merchant
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/PathID' }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateMerchantRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MerchantResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/merchants/{id}/status:
    put:
      tags: [Admin - Merchants]
      summary: Activate/deactivate a merchant
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/PathID' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { is_active: { type: boolean } } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MerchantResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin session token from POST /auth/admin/login.
    merchantBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Merchant dashboard session token from POST /auth/login.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Merchant API key. Accepted either as `X-API-Key: <key>` or
        `Authorization: Bearer <key>` (see MerchantAPIAuthMiddleware).
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }

    # ---- Auth ----
...
    CreateMerchantRequest:
      type: object
      required: [name, email, business_name]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        business_name: { type: string }
        webhook_url: { type: string, nullable: true }
    UpdateMerchantRequest:
      type: object
      properties:
        name: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        business_name: { type: string, nullable: true }
        webhook_url: { type: string, nullable: true }
    MerchantResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        business_name: { type: string }
        webhook_url: { type: string, nullable: true }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PaginatedMerchants:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/MerchantResponse' } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }

    # ---- API Keys ----
...